L3 layout: _inGroup int* → byte* (58KB → 14KB; +3% avg bench)
The _inGroup dedup bitmap held 0/1 only — int* was 4× wasteful. Switched
to byte* so the dedup array fits L1d alongside NodeStates (both now 14KB,
total 28KB of the 32KB typical L1d).

Bit-identical (checksum 0x933ABE7915AC18BE on full_palette.nes 50K hc).

A/B bench (full_palette.nes 50K hc, 5 runs each, same machine state):
  baseline (int*):  best 37,928  avg 36,923  hc/s
  modified (byte*): best 38,534  avg 38,018  hc/s
                    → +1.6% best, +3.0% avg

Reads (`_inGroup[nn] != 0`) and writes (`_inGroup[nn] = 1`) work
identically for byte* — 0/1 values fit either. No code logic changed,
only the type declaration and the AllocArray<T> width.

Part of L3 layout investigation (size each hot array, find where cache
behavior matters). Next candidates by size: TransistorList int→short
(697KB → 350KB), NodeInfos struct SoA-split, RecalcHash int→bit.

+++ src/AprVisual/Sim/WireCore.Group.cs
@@ -28,7 +28,7 @@ namespace AprVisual.Sim
         private static NodeFlags _groupFlags;
         private static int _groupCount;       // number of nodes currently in _groupBuf
         private static int* _groupBuf;        // node ids in the current group (alloc'd in Reset, sized NodeCount)
-        private static int* _inGroup;         // O(1) dedup flag per node (1 = currently in _groupBuf); cleared each ComputeNodeGroup
+        private static byte* _inGroup;        // O(1) dedup flag per node (1 = currently in _groupBuf); cleared each ComputeNodeGroup. byte* (was int*) — 0/1 only, 14KB vs 58KB fits L1d alongside NodeStates
         private static byte _maxState;        // state of the highest-connection node seen
         private static int _maxConnections;
